+++ src/hooks/useValidation/rules.js
@@ -90,26 +90,7 @@ const BASE_BIB_VALIDATORS = [
   },
   {
     validator: RULES.CONTROLLED_SUBFIELD,
-    message: (uniqueTags) => {
-      if (uniqueTags.length === 1) {
-        return ({
-          id: 'ui-quick-marc.record.error.fieldIsControlled',
-          values: {
-            count: 1,
-            fieldTags: `MARC ${uniqueTags[0]}`,
-          },
-        });
-      }
-
-      return ({
-        id: 'ui-quick-marc.record.error.fieldsAreControlled',
-        values: {
-          count: uniqueTags.length,
-          fieldTags: uniqueTags.slice(0, -1).map(tag => `MARC ${tag}`).join(', '),
-          lastFieldTag: `MARC ${uniqueTags[uniqueTags.length - 1]}`,
-        },
-      });
-    },
+    message: () => ({ id: 'ui-quick-marc.record.error.fieldIsControlled' }),
   },
   {
     tag: '010',